Set an appropriate image size for the screenshots
Display the screenshots in a moderate size which doesn't interrupt the reading flow that much. Note that GFM doesn't provide any native image sizing capabilities, so we have to fall back on good old HTML. I'm using the relative unit `em` so that the images will be displayed nicely on high DPI environments, too.